pyproject.toml
[tool.poetry]
name = "banking-transactions"
version = "0.1.0"
description = ""
authors = ["kirill_tarasov <kirilltaras1792@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
datetime = "^5.3"
pandas = "^2.1.3"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
mypy = "^1.7.1"
black = "^23.11.0"
isort = "^5.12.0"

[tool.black]
line-length = 119

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
